Benchmark data generator created.
The modules `names` and `words` are just helpers to generate random
names and text, respectively.

The module `dataset` has the code to generate a random movie dataset
roughly with the same properties as real-life: the ratio or actors to
directors, how many movies directors make and how many actors are
reported as "starring" in those.

The module `edgedb.importer` has a simple function that takes a
DataGenerator and produces EQL as output.