Update the benchmarks.
Replace the data generator to use `synth` and produce a couple of JSON
files (one that is more object-oriented and the other more
table-oriented).

Conver the EdgeDB benchmark to use `edgedb project` functionality.
Update how the different modes of querying are accessed.

Update the EdgeDB drivers used in other benchmarks.

Update the other libraries being benchmarked to their respective current
versions.

Drop Loopback benchmark due to the fact that the old ones used a stale
version (Loopback 3) and the new Loopback 4 changed how the projects are
setup. So it would take more time to re-learn, while there's no reason
(based on the past performance) to believe that it's particularly
performant. Whoever is using Loopback is probably using it for reasons
that are more important to them than the fairly slow ORM component.