Add an "update" test to the benchmarks. (#30)
* Add an "update" and "insert" test to the benchmarks.

The update is very basic. The assumption is that a lot of the updates
are driven by forms and therefore retrieving the data to be updated is
not necessarily part of the update call. It is also assumed that many
updates modify very few fields at a time: updating email, changing
status, etc. So the relevant benchmark would be trying to update a field
on an object, without nesting or other complexity.

The insert is also similarly basic. An object without exclusive fields
is used in order to avoid conflicts.

* Add nested inserts.

Postgraphile is currently missing nested inserts out-of-the-box.

All other benchmarks have the nested inserts implemented based on
inserting a Movie.