set expire_on_commit=False
The expire_on_commit default behavior is intended to
allow the ORM Session to work consistently across
transaction commits in an environment where the state of the
database is assumed to be changing concurrently, including
against data that was just locally inserted.

The implication of expire_on_commit is that objects which
were just committed will be fully expired, meaning any access
to their attributes subsequent will incur new queries
to "refresh" the data.   This is in a wide range of cases
unnecessary and wasteful from a performance perspective,
not to mention that it gets in the way of asyncio
patterns where lazy loading is not an option.

For web services that are INSERTing new data, the action
to retrieve that data subsequent to COMMIT should in
usual cases not return expired objects, as the method
seeks to return the information that is exactly
what was just inserted.  So here we turn this parameter
off since it definitely will otherwise negatively
impact measured performance.