use asyncpg as the driver
The psycopg2 driver is SQLAlchemy's default driver
for PostgreSQL, however is now obsolete as the psycopg3
project is now released, including
lots of modernizations and performance improvements as well
as async support.  SQLAlchemy's support for psycopg3 is
part of not-released SQLAlchemy 2.0.  In the interim, we can
use asyncpg for all features here including sync APIs by
setting up async_fallback=True.  Establishing asyncpg driver
support here sets us up to also add a queries_asyncio module
in an upcoming commit.

The move to asyncpg seems to necessitate that we use a
"TIMESTAMP WITH TIMEZONE" datatype as the data coming from the .json
appears to include Python timezones, so add the timezone
parameter to the one DateTime column.