update models and queries
in this commit we use modern APIs and features for the
ORM load/commit demonstration:

1. use select() for selecting; the "baked queries" extension
   is now legacy as SQLAlchemy's query caching mechanism has
   been generalized to all queries.  A "baked" still has a slight
   per-query edge when the "lambda" feature is used,
   which we could have used here as well without "baked",
   but I would prefer this code illustrate clean and idiomatic
   API use rather than more esoteric speed improvements.

2. bulk updates/deletes use update(), delete()

3. remove use of subqueryload() which is superseded by
   selectinload(), which generates much simpler queries that
   should generally plan better

4. use a with_parent() query instead of lazy="dynamic", this
   relationship style isn't going away but is a little up in
   the air right now

5. set innerjoin=True on relationships that are M2O against
   non-nullable foreign key col

The performance here is not necessarily expected to improve
or worsen, factors which should improve performance include:

1. setting expire_on_commit=False
2. using selectinload()

Factors which might degrade performance include:

1. a baked with lambdas is slightly faster than a regular
   statement that isn't built up with lambdas, as the generation
   of the construct as well as the cache key is much faster
   with lambdas.   We hope to increase the use of cython
   in the 2.0 series to optimize cache key generation which hopefully
   will help alot.

Performance is still not optimal for what we can do on the
persistence side, the ORM now knows how to INSERT..RETURNING
multiple records in one statement, which is a *huge* performance
increase over statement-at-a-time, but in 1.4 this is only
implemented for psycopg2 so far and at the moment the revised
demo is using asyncpg.  As the multi-insert-returning
demo here seems to involve only a few records per commit,
stick with the old way.